Added new Java-script for generating AI file
Java-script need to be run in Adobe Illustrator, and generates a braille like file (circles instead of dots) with standard braille dimentions.
The file is to laser cut it on acrilyc sheet.
Laser power and feedrate should be adjusted so it cuts a holes with 0.5-0.7 millimitres in depth.
After that acrilyc sheet will be used as a mold for silicon pouring.